Add -pollaxes flag for special case joysticks.
For most joysticks we can detect when a stick has been pushed outside
of the dead zone towards an extreme. But when testing the Gravis
Gamepad with a Gamepoert to USB adapter, I found that it was not
possible to register a value over ~20000 in some directions. In this
kind of case it is not possible using the event code to tell when
the stick has been pushed in those directions. So add -pollaxes that
will poll once a second and print the value of each axis, as a fallback
in case the usual event code doesn't work.

Also, adjust the dead zone limit so that > 1/3 of the full range is
considered pushing to an extreme. This is consistent with the Chocolate
Doom joystick code that assumes this.